a function that prints an integer

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -9,7 +9,7 @@
   */
 void print_number(int n)
 {
-long int x = n;
+unsigned int x = n;
 
 if (n < 0)
 {
